Only define extruder, heater, and thermistor pins for drivers on board
In case of manual switching toolhead the number of hotends generally exceeds the number of extruder drivers and the number heater and thermistor pins on the board. Thus in this case, we need to limit the definition of the driver, heater, thermistor pins to those actually available on the board.